+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to the "veracode" extension will be documented in this file.
 
+## [0.0.5]
+- Rewrite the code to add Types
+- Upgrade dependencies to the latest vscode
+- Added ability to read credentials from different sections in the credentials file
+- Added ability to view multiple issues in the 'PROBLEMS' pane (in oppose to one at a time)
+- Added configuration file to filter applications and sandbox in the veracode view
+
 ## [0.0.4]
 - download only
 - significant internal re-work due to the fact that extensins run differently in the debugger vs. a normal install